Extract communitcation functions to a separate file
